add db-check
